+++ application/views/comments/home.php
@@ -1,8 +1,8 @@
 <ul>
   <li>
-    <a href='<?php echo BRANDURL; ?>articles/write/'>Write Article</a>
+    <a href='<?php echo DOMAINURL; ?>articles/write/'>Write Article</a>
   </li>
   <li>
-    <a href='<?php echo BRANDURL; ?>articles/viewall/'>View All Articles</a>
+    <a href='<?php echo DOMAINURL; ?>articles/viewall/'>View All Articles</a>
   </li>
 </ul>