+++ application/views/domains/create.php
@@ -1,0 +1,10 @@
+<!-- Results -->
+<div id='results'>
+  <?php
+  if (isset($domainID)
+    && $domainID != 'DEFAULT') {
+    echo "Brand (#".$domainID.") updated!";
+  }
+  ?>
+</div>
+